requirements.txt
# Client Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
paho-mqtt>=1.6.1
PyYAML>=6.0
python-dotenv>=1.0.0

# Sensor libraries (for Raspberry Pi)
adafruit-circuitpython-dht>=3.7.0
board>=1.0
w1thermsensor>=2.0.0

# Modbus support
pymodbus>=3.1.0

# Additional utilities
requests>=2.28.0
schedule>=1.2.0
psutil>=5.9.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
coverage>=7.0.0

# Optional: For web debug console
flask>=2.3.0
flask-socketio>=5.3.0